fix: split waste calendar generator to satisfy complexity gate
Agent-Logs-Url: https://github.com/smart-village-solutions/sva-studio/sessions/b694dde6-71bd-4a5a-b20b-4a0e7af5dcf8
